Module Mrmime__Header

type t

Type of the header.

val pp : t Fmt.t

Pretty-printer of pp.

val assoc : Mrmime.Field_name.t -> t -> Mrmime.Field.field list

assoc field_name header returns all values associated to field_name (several can exist into a header). It returns order, value and location if it comes from a source.

val remove_assoc : Mrmime.Field_name.t -> t -> t

remove_assoc field_name t removes all values associated to field_name.

val concat : t -> t -> t

concat t0 t1 concats t0 with t1 - any field-name existing in both are not deleted/replaced.

val exists : Mrmime.Field_name.t -> t -> bool

exists field_name t is true if field_name exists in t.

val add : Mrmime.Field_name.t -> ('a Mrmime.Field.t * 'a) -> t -> t

add field_name (w, v) t adds a new field-name with value v. add does not replace field_name if it already exists into t.

val replace : Mrmime.Field_name.t -> ('a Mrmime.Field.t * 'a) -> t -> t

replace field_name (w, v) t replaces existing field-name field_name in t by the new value v. If field_name does not exist, it adds it.

val of_list : Mrmime.Field.field list -> t

of_list l returns a header from the list l (without location).

val of_list_with_location : Mrmime.Field.field Mrmime.Location.with_location list -> t

val empty : t

empty is an empty header which does not have any default values.

val content_type : t -> Mrmime.Content_type.t

content_type header returns Content_type.t of the given header. If it does not exist, it returns Content_type.default.

val content_encoding : t -> Mrmime.Content_encoding.t

content_encoding header returns Content_encoding.t of the given header. If it does not exist, it returns Content_encoding.default.

Decoder of header.

module Decoder : sig ... end

Encoder of header.

module Encoder : sig ... end

val to_stream : t -> unit -> string option

to_stream header returns a stream of the given header which can be used into protocol like SMTP.