Module Mrmime__Mailbox.Domain

type atom = [

| `Atom of string

]

type literal = [

| `Literal of string

]

type 'a domain =

| (::) : atom * 'a domain -> 'a Mrmime__.Peano.s domain

| ([]) : Mrmime__.Peano.z domain

type 'a t

Kind of domain. RFC 5322 and RFC 5321 allows several kinds of domain:

An usual domain which is a non-empty list of atom elements

A Literal_domain.t

A `Literal domain which is a string surrounded by brackets.

val atom : string -> (atom, [> Rresult.R.msg ]) Stdlib.result

atom x returns a safe atom element. If x does not respect RFC 5322, it returns Error. It accepts any characters excepts controls, space and specials characters - for instance, brackets are not allowed.

val atom_exn : string -> atom

Same as atom but raises an Invalid_argument instead Error.

val a : string -> atom

Alias of atom_exn.

val literal : string -> (literal, [> Rresult.R.msg ]) Stdlib.result

literal x returns a literal domain. If x does not respect RFC 5321, it returns Error. It will try to escape control characters (with escape_string).

val literal_exn : string -> literal

Same as literal but raises an Invalid_argument instead to return Error.

val domain : 'a domain t

Kind of domain.

val ipv4 : Ipaddr.V4.t t

Kind of Literal_domain.ipv4.

val ipv6 : Ipaddr.V6.t t

Kind of Literal_domain.ipv6.

val extension : (string * string) t

Kind of Literal_domain.extension.

val default : string t

Kind of literal.

val make : 'a t -> 'a -> (Emile.domain, [> Rresult.R.msg ]) Stdlib.result

make kind v returns a safe domain. It can fail if an user-defined literal-domain (Literal_domain.extension), a literal domain or a domain don't follow standards:

for a Literal_domain.extension, make returns Error if Literal_domain.make returns Error

for a literal, make returns Error if literal returns Error

for a domain, make returns Error if list of atom is empty

val of_list : string list -> (Emile.domain, [> Rresult.R.msg ]) Stdlib.result

of_list l returns a domain from a non-empty list of well-formed atom elements. Otherwise, it returns an error.

val v : 'a t -> 'a -> Emile.domain

Same as make but raises an Invalid_argument instead Error.

val to_string : Emile.domain -> string

to_string x returns a string which represents x as is it in a e-mails.