Module Prettym__.Enclosure

module Buffer : sig ... end

module IOVec : sig ... end

type encoder

The type of encoder.

val is_empty : encoder -> bool

is_empty t returns true if nothing is under t. This case appear afterwards a flush.

val pp : encoder Fmt.t

Pretty-printer of encoder.

val schedule_flush : (int -> encoder -> unit) -> encoder -> encoder

schedule_flush f encoder registers f to be called when all prior writes have been successfully completed. If encoder has no pending writes, then f will be called immediately.

val schedule_bigstring : encoder -> ?⁠off:int -> ?⁠len:int -> Bigstringaf.t -> encoder

schedule_bigstring t ?off ?len a stores a pointer to a into the serializer's internal queue.

val kschedule_bigstring : (encoder -> 'r) -> encoder -> ?⁠off:int -> ?⁠len:int -> Bigstringaf.t -> 'r

kschedule_bigstring k t ?off ?len a: kontinuation-style of schedule_bigstring.

val schedule_string : encoder -> ?⁠off:int -> ?⁠len:int -> string -> encoder

Same as schedule_bigstring but for String.t.

val kschedule_string : (encoder -> 'r) -> encoder -> ?⁠off:int -> ?⁠len:int -> Stdlib.String.t -> 'r

kschedule_string k t ?off ?len a: kontinuation-style of schedule_string.

val schedule_bytes : encoder -> ?⁠off:int -> ?⁠len:int -> bytes -> encoder

Same as schedule_bigstring but for Bytes.t.

val kschedule_bytes : (encoder -> 'r) -> encoder -> ?⁠off:int -> ?⁠len:int -> Stdlib.Bytes.t -> 'r

kschedule_bytes k t ?off ?len a: kontinuation-style of schedule_bytes.

val write_char : char -> encoder -> encoder

write_char chr t copies chr into the serializer's internal buffer.

val write_string : ?⁠off:int -> ?⁠len:int -> string -> encoder -> encoder

write_string ?off ?len x t copies x into the serializer's internal buffer.

val write_bigstring : ?⁠off:int -> ?⁠len:int -> Bigstringaf.t -> encoder -> encoder

Same as write_string but for bigstring.

val write_bytes : ?⁠off:int -> ?⁠len:int -> bytes -> encoder -> encoder

Same as write_string but for Bytes.t.

val write_uint8 : int -> encoder -> encoder

write_uint8 x t copies the lower 8 bits of x into the serializer's internal buffer.

val flush : (encoder -> 'r) -> encoder -> 'r

flush k encoder enforces to flush at this stage encoder. Then, when end-user flushes encoder, k will be called with the new state of encoder drained.

module LE : sig ... end

module BE : sig ... end

type emitter = IOVec.t list -> int

val create : emitter:emitter -> int -> encoder

create ~emitter len creates a serializer with a fixed-length internal buffer. When internal buffers are full, encoder calls emitter to emit IOVec.t list.