Module Date.Zone

type t =

| UT

| GMT

| EST

| EDT

| CST

| CDT

| MST

| MDT

| PST

| PDT

| Military_zone of char

| TZ of int * int
(*
Type of zone according RFC 822 / RFC 2822 / RFC 5322.
*)

Constructors.

val ut : t

val gmt : t

val est : t

val edt : t

val cst : t

val cdt : t

val mst : t

val mdt : t

val pst : t

val pdt : t

val military_zone : char -> (t, [ `Msg of string ]) Stdlib.result

val tz : int -> int -> (t, [ `Msg of string ]) Stdlib.result

val to_string : t -> string

to_string v returns a well-formed string from a time zone v.

val of_string : string -> (t, [ `Msg of string ]) Stdlib.result

of_string v returns a time zone from a well-formed string v. Process is case-sensitive.

val of_string_exn : string -> t

of_string_exn v returns a time zone from a well-formed string v. Process is case-sensitive.

raises [Invalid_argument]
when v is an invalid time zone.

val v : string -> t

Alias of of_string_exn.

Pretty-printers.

val pp : Stdlib.Format.formatter -> t -> unit

Equals.

val equal : t -> t -> bool