Module Mrmime.Group

type t = Emile.group

Type of group of addresses.

module Phrase : sig ... end

val make : name:Emile.phrase -> Mailbox.t list -> t option

make ~name mailboxes makes a new group with name name of mailboxes.

val v : name:Emile.phrase -> Mailbox.t list -> t

Same as make but raises an exception if list of mailboxes is empty.

Equal.

val equal : t -> t -> bool

Equal function of t.

Pretty-printer.

val pp : Stdlib.Format.formatter -> t -> unit

Pretty-printer of t.

Decoder of group.

module Decoder : sig ... end

Encoder of group.

module Encoder : sig ... end