Module Mailbox.Literal_domain

type 'a t

A literal-domain type. Sometimes a host is not known to the domain name system and communication (and, in particular, communication to report and repair the error) is blocked. To bypass this barrier, a special literal form of the address is allowed as an alternative to a domain name.

We have three kinds of literal-domain:

an IPv4 address

an IPv6 address (textual form follows RFC 4291)

an user-defined literal domain

val ipv4 : Ipaddr.V4.t t

IPv4 kind.

val ipv6 : Ipaddr.V6.t t

IPv6 kind.

val extension : (string * string) t

An user-defined literal domain kind.

val make : 'a t -> 'a -> (Emile.addr, [> `Msg of string ]) Stdlib.result

make kind v returns a literal-domain according RFC 5321. It should fails if kind is extension and value does not respect standards.

val v : 'a t -> 'a -> Emile.addr

Same as make but raises an exception instead to return Error.